Use DateTime instead of Time for timestamps
Now that we're using redis as caching service, values are serialised
with Marshal, which has issues with Time instances too old in the past
and our Ruby version (2.4), so errors like:

```
ArgumentError (year too big to marshal: 1800 UTC)
```

are raised when a cequel model instance with an old timestamp is written
via redis-store (Check https://bugs.ruby-lang.org/issues/15160). To
solve the problem, the places where `Time` is used have been changed to
`DateTime` instances.